requirements.txt
requests>=2.20.0
beautifulsoup4>=4.6.0
lxml>=4.0.0
postmarker>=1.0.0
python-dotenv>=0.10.0
